Triangles' raytracing added.
Two methods for testing that a point P is inside the triangle are prov$
a) "Same side" method:
Compares P's position to every edge of the triangle, using the non-edge
vertex, in order to determine if it is in the right side of the edge in
order to be inside of the triangle.

b) Barycentric coordinates system methods:
Calculates the barycentric coordinates (u,v) of the point P, taking A
vertex as the origin of coordinates.

The point P is inside the triangle iif:
- 0 <= u <= 1
- 0 <= v <= 1
- u + v <= 1